+++ dashboard/pages/1.0 sort.ps1
@@ -1373,7 +1373,7 @@ New-UDPage -Name "Sort" -Content {
                                     Show-UDModal -FullWidth -MaxWidth xl -Content {
                                         $scrapers = @(
                                             'aventertainment',
-                                            'aventertainmentja'
+                                            'aventertainmentja',
                                             'dmm',
                                             'dmmja',
                                             'jav321ja',
@@ -1384,8 +1384,6 @@ New-UDPage -Name "Sort" -Content {
                                             'javlibraryja',
                                             'javlibraryzh',
                                             'mgstageja'
-                                            'r18',
-                                            'r18zh'
                                         )
                                         New-UDTextbox -Id 'textbox-sort-manualsearch' -Placeholder 'Enter an ID or a comma separated list of URLs' -FullWidth -MultiLine -Autofocus
                                         New-UDGrid -Container -Content {
@@ -1420,8 +1418,6 @@ New-UDPage -Name "Sort" -Content {
                                                         Javlibrary   = $cache:settings.'web.sort.manualsearch.javlibrary'
                                                         JavlibraryJa = $cache:settings.'web.sort.manualsearch.javlibraryja'
                                                         JavlibraryZh = $cache:settings.'web.sort.manualsearch.javlibraryzh'
-                                                        R18          = $cache:settings.'web.sort.manualsearch.r18'
-                                                        R18Zh        = $cache:settings.'web.sort.manualsearch.r18zh'
                                                         Aggregated   = $true
                                                     }
                                                     $jvData = (Javinizer @findParams)